Move client to replica/client and controller/client
