+++ controller/rest/router.go
@@ -1,26 +1,15 @@
 package rest
 
 import (
-	"net/http"
-
 	"github.com/gorilla/mux"
 	"github.com/rancher/go-rancher/api"
-	"github.com/rancher/go-rancher/client"
+	"github.com/rancher/longhorn/replica/rest"
 )
-
-func HandleError(s *client.Schemas, t func(http.ResponseWriter, *http.Request) error) http.Handler {
-	return api.ApiHandler(s, http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		if err := t(rw, req); err != nil {
-			apiContext := api.GetApiContext(req)
-			apiContext.WriteErr(err)
-		}
-	}))
-}
 
 func NewRouter(s *Server) *mux.Router {
 	schemas := NewSchema()
 	router := mux.NewRouter().StrictSlash(true)
-	f := HandleError
+	f := rest.HandleError
 
 	// API framework routes
 	router.Methods("GET").Path("/").Handler(api.VersionsHandler(schemas, "v1"))
